Support draw calls for subpixel mouse coordinates
Also drop the overscale since it doesn't appear to be required anymore.

+++ CorsixTH/Src/th_gfx_font.cpp
@@ -80,7 +80,7 @@ text_layout bitmap_font::get_text_dimensions(const char* sMessage,
 }
 
 void bitmap_font::draw_text(render_target* pCanvas, const char* sMessage,
-                            size_t iMessageLength, int iX, int iY) const {
+                            size_t iMessageLength, float iX, float iY) const {
   pCanvas->start_nonoverlapping_draws();
   if (iMessageLength != 0 && sheet != nullptr) {
     const unsigned int iFirstASCII = 31;
@@ -101,7 +101,7 @@ void bitmap_font::draw_text(render_target* pCanvas, const char* sMessage,
         sheet->get_sprite_size_unchecked(iChar, &iWidth, &iHeight);
         iWidth *= scale_factor;
         iHeight *= scale_factor;
-        iX += iWidth + scaled_letter_spacing;
+        iX += static_cast<float>(iWidth + scaled_letter_spacing);
       }
     }
   }
@@ -110,8 +110,8 @@ void bitmap_font::draw_text(render_target* pCanvas, const char* sMessage,
 
 text_layout bitmap_font::draw_text_wrapped(render_target* pCanvas,
                                            const char* sMessage,
-                                           size_t iMessageLength, int iX,
-                                           int iY, int iWidth, int iMaxRows,
+                                           size_t iMessageLength, float iX,
+                                           float iY, int iWidth, int iMaxRows,
                                            int iSkipRows,
                                            text_alignment eAlign) const {
   text_layout oDrawArea = {};
@@ -179,18 +179,18 @@ text_layout bitmap_font::draw_text_wrapped(render_target* pCanvas,
           draw_text(pCanvas, sMessage, sBreakPosition - sMessage, iX + iXOffset,
                     iY);
         }
-        iY += static_cast<int>(iTallest) + scaled_line_spacing;
+        iY += static_cast<float>(iTallest + scaled_line_spacing);
         oDrawArea.end_x = iMsgWidth;
         oDrawArea.row_count++;
         if (foundNewLine) {
-          iY += static_cast<int>(iTallest) + scaled_line_spacing;
+          iY += static_cast<float>(iTallest + scaled_line_spacing);
           oDrawArea.row_count++;
         }
       } else {
         iSkippedRows++;
         if (foundNewLine) {
           if (iSkippedRows == iSkipRows) {
-            iY += static_cast<int>(iTallest) + scaled_line_spacing;
+            iY += static_cast<float>(iTallest + scaled_line_spacing);
             oDrawArea.row_count++;
           }
           iSkippedRows++;
@@ -375,7 +375,7 @@ text_layout freetype_font::get_text_dimensions(const char* sMessage,
 }
 
 void freetype_font::draw_text(render_target* pCanvas, const char* sMessage,
-                              size_t iMessageLength, int iX, int iY) const {
+                              size_t iMessageLength, float iX, float iY) const {
   draw_text_wrapped(pCanvas, sMessage, iMessageLength, iX, iY, INT_MAX);
 }
 
@@ -415,8 +415,8 @@ FT_Pos pixel_align(FT_Pos position) { return ((position + 63) >> 6) << 6; }
 
 text_layout freetype_font::draw_text_wrapped(render_target* pCanvas,
                                              const char* sMessage,
-                                             size_t iMessageLength, int iX,
-                                             int iY, int iWidth, int iMaxRows,
+                                             size_t iMessageLength, float iX,
+                                             float iY, int iWidth, int iMaxRows,
                                              int iSkipRows,
                                              text_alignment eAlign) const {
   text_layout oDrawArea = {};
@@ -724,8 +724,8 @@ text_layout freetype_font::draw_text_wrapped(render_target* pCanvas,
     draw_texture(pCanvas, pEntry, iX, iY);
   }
   oDrawArea.width = pEntry->widest_line_width;
-  oDrawArea.end_x = iX + pEntry->last_x;
-  oDrawArea.end_y = iY + pEntry->height;
+  oDrawArea.end_x = static_cast<int>(iX) + pEntry->last_x;
+  oDrawArea.end_y = static_cast<int>(iY) + pEntry->height;
   oDrawArea.row_count = pEntry->row_count;
   return oDrawArea;
 }
